+++ test/add_test_columns_migration.rb
@@ -1,7 +1,7 @@
 class AddTestColumns < ActiveRecord::Migration[5.2]
   def change
     add_column :policy_elements, :color, :string
-    if ActiveRecord::Base.connection.class.name == 'ActiveRecord::ConnectionAdapters::PostgreSQLAdapter'
+    if PolicyMachineStorageAdapter.postgres?
       add_column :policy_elements, :tags, 'text[]'
       add_column :policy_elements, :document, :jsonb, default: {}
     else